+++ include/coal/narrowphase/support_functions.h
@@ -289,8 +289,8 @@ void getShapeSupportSet(const LargeConvex<IndexType>* convex,
 /// @param[in] cloud data which contains the 2d points of the support set which
 /// convex-hull we want to compute.
 /// @param[out] 2d points of the the support set's convex-hull.
-COAL_DLLAPI void computeSupportSetConvexHull(SupportSet::Polygon& cloud,
-                                             SupportSet::Polygon& cvx_hull);
+COAL_DLLAPI void computeSupportSetConvexHull(const std::vector<Vec2s>& cloud,
+                                             std::vector<Vec2s>& cvx_hull);
 
 }  // namespace details
 }  // namespace coal