+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add pixi-build support ([#774](https://github.com/coal-library/coal/pull/774))
 - contact patches: add helpers to reduce the size of a contact patch by keeping points that minimize the difference between the area of the original patch and the area of the new patch
   - handle the specific cases of n_desired = 1 or 2 ([805](https://github.com/coal-library/coal/pull/805))
+  - improve efficiency of ContactPatchSimplifierMaxArea and improve testing of patch simplifiers ([813](https://github.com/coal-library/coal/pull/813))
 - timings: + and += operators ([805](https://github.com/coal-library/coal/pull/805))
 
 ### Removed